Move Region into its own file
