+++ crates/bitwarden-importers/src/importers/onepassword/access/mod.rs
@@ -8,7 +8,7 @@ mod account_key;
 mod client;
 pub use client::Client;
 mod credentials;
-pub use credentials::{Credentials, Region};
+pub use credentials::Credentials;
 mod device;
 pub use device::generate_device_uuid;
 mod error;
@@ -21,6 +21,8 @@ mod mac;
 pub mod model;
 pub use model::{Item, ItemCategory, Vault};
 mod opdata;
+mod region;
+pub use region::Region;
 mod rest;
 mod rsa;
 mod session;
